Update date display to show full month names
Modify date formatting across multiple components to use "Month Day, Year" instead of "Mon Day, Year".

+++ artifacts/mobile/app/insurance-quote.tsx
@@ -48,7 +48,7 @@ function fmt(iso: string): string {
   if (!iso) return "";
   const d = new Date(iso);
   if (isNaN(d.getTime())) return iso;
-  return d.toLocaleDateString("en-US", { month: "short", day: "numeric", year: "numeric" });
+  return d.toLocaleDateString("en-US", { month: "long", day: "numeric", year: "numeric" });
 }
 
 export default function InsuranceQuoteScreen() {
