+++ artifacts/mobile/app/piece/[id].tsx
@@ -72,7 +72,7 @@ function formatDate(iso: string): string {
   if (!iso) return "—";
   const d = new Date(iso);
   if (isNaN(d.getTime())) return iso;
-  return d.toLocaleDateString("en-US", { month: "short", day: "numeric", year: "numeric" });
+  return d.toLocaleDateString("en-US", { month: "long", day: "numeric", year: "numeric" });
 }
 
 function daysStatus(expiry: string): { text: string; color: string } | null {
